Visa npc i grupp och radera npc - admin

+++ admin/logic/delete_npc.php
@@ -5,21 +5,21 @@ if ($_SERVER["REQUEST_METHOD"] == "GET") {
     
     if (isset($_GET['id'])) {
         $npc = NPC::loadById($_GET['id']);
+        if (isset($npc)) NPC::delete($npc->Id);
+    } elseif (isset($_GET['roleID'])) {
+        $role = Role::loadById($_GET['roleID']);
+        if (isset($role) && $role->mayDelete()) Role::delete($role->Id);
     } else {
-        
         header('Location: ../index.php');
         exit;
     }
 }
 
-
-if (!isset($npc)) {
-    header('Location: ../npc.php');
+if (isset($_SERVER['HTTP_REFERER'])) {
+    header('Location: ' . $_SERVER['HTTP_REFERER']);
     exit;
 }
 
-
-NPC::delete($npc->Id);
 
 header('Location: ../npc.php');
